requirements.txt
torch
torchvision
opencv-python==4.5.4.60
numpy>=1.21
Pillow>=9.1
tqdm>=4